requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
gunicorn==23.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
packaging==24.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.2
sympy==1.13.2
torch==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
